MVP complete!  classes change on task items and the todoList in state is updated so that taskObj.completed === true after click and toggles back to false if clicked again.  Clear completed button will only prompt an 'Are you sure you want to...' message if there has been one or more tasks completed.  If not, it does nothing.  If clear completed button is clicked and user picks ok to clear completed, the todoList in state is updated with the new list, which removed the completed items/tasks.  May work on search functionality in stretch.
