+++ src/components/Todo.css
@@ -1,7 +1,7 @@
 .Todo-container {
     border: 1px solid black;
     box-shadow: 3px 3px 3px black;
-    width: 25%;
+    width: 15%;
     margin: 2% 0 2%;
 }
 
@@ -11,6 +11,7 @@
 
 .Todo-container .todo-item {
     text-decoration: none;
+    cursor: pointer;
 }
 
 .Todo-container .completed {
