+++ src/dummyData.js
@@ -1,0 +1,14 @@
+const dummyData = [
+    {
+      task: 'Organize Garage',
+      id: Date.now()*100,
+      completed: false
+    },
+    {
+      task: 'Bake Cookies',
+      id: Date.now()*50,
+      completed: false
+    },
+];
+
+export default dummyData;
